refactor: move logic to a seperate patch
